api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.12
passlib[argon2]==1.7.4
python-jose==3.3.0
itsdangerous==2.2.0
httpx==0.27.2
redis==5.0.8
